Fix update_system's on_change on Redhat
Redhat can have multiple packages installed with the same name (eg
kernels), a fact that Rex::Pkg::Base::diff_package_list is not
prepared to deal with, resulting in calling the on_change hook when
you run the update_system command, even if nothing changed in the
packages installed (see also bug #1149). This commit fixes that.